Remove .ruby-version for development
RubyGems may be used across a wide range of Ruby versions,
from the minimum supported version to development versions.

To allow local development with any Ruby version,
the Ruby version is not locked using a `.ruby-version` file.

This enables developers to work with their preferred Ruby version and
frees repository maintainers from the need to update `.ruby-version`.

Additionally, because testing against supported Ruby versions is performed
in CI before release, the version in `.ruby-version` is not important.